requirement.txt
asgiref==3.9.1
Django==4.2.24
django-filter==25.1
djangorestframework==3.16.1
dotenv==0.9.9
importlib_metadata==8.7.0
Markdown==3.9
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
zipp==3.23.0
